refactoring code, resolve more conflicts
